+++ src/Products/PloneMeeting/ToolPloneMeeting.py
@@ -1315,7 +1315,7 @@ class ToolPloneMeeting(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
            - If p_prefix is True, the translated prefix "Meeting of" is
              prepended to the result.'''
         # Received meeting could be a brain or an object
-        if meeting.__class__.__name__ in ['mybrains', 'CatalogContentListingObject']:
+        if meeting.__class__.__name__ in ['mybrains', 'CatalogContentListingObject', 'PloneFlare']:
             # It is a meeting brain, take the 'getDate' metadata
             date = meeting.getDate
         else: