+++ src/Products/PloneMeeting/config.py
@@ -33,6 +33,12 @@ try:
 except pkg_resources.DistributionNotFound:
     HAS_ZAMQP = False
 
+HAS_SOLR = True
+try:
+    pkg_resources.get_distribution('collective.solr')
+except pkg_resources.DistributionNotFound:
+    HAS_SOLR = False
+
 __author__ = """Gaetan DELANNAY <gaetan.delannay@geezteem.com>, Gauthier BASTIEN
 <g.bastien@imio.be>, Stephan GEULETTE <s.geulette@imio.be>"""
 __docformat__ = 'plaintext'