+++ src/Products/PloneMeeting/tests/testMeetingItem.py
@@ -4284,7 +4284,10 @@ class testMeetingItem(PloneMeetingTestCase):
                 cloned_to_same_mc=True, cloned_from_item_template=False) +
             item.adapted().getExtraFieldsToCopyWhenCloning(
                 cloned_to_same_mc=False, cloned_from_item_template=False))
-        self.assertEquals(copiedFields, set(itemFields))
+        # showinsearch and searchwords must be ignored when using Solr
+        item_field_set = set(
+                [field_name for field_name in itemFields if field_name not in ('showinsearch', 'searchwords')])
+        self.assertEquals(copiedFields, item_field_set)
 
         newItem = item.clone()
         self.assertEquals(item.Title(), newItem.Title())