+++ CHANGES.rst
@@ -32,6 +32,7 @@ Changelog
 - Moved MeetingConfig.onMeetingTransitionItemTransitionToTrigger to MeetingConfig.onMeetingTransitionItemActionToExecute, in addition to be able to trigger a transition on every items
   of a meeting when a transition is triggered on a meeting, it is now possible to execute a TAL expression
 - 'workflowstate' viewlet was moved to plonetheme.imioapps.browser.viewlets and utils.get_state_infos was moved to imio.helpers.content, adapted code accordingly
+- Added Ability to run using solr instead of catalog
 
 4.1 (2019-08-23)
 ----------------